Rework video player component to with-video HOC
